pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tinygo"
version = "0.5.0"
description = "CLI tool for deploying web pages to tiiny.host"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.28",
    "rich>=13.0",
]

[project.optional-dependencies]
aws = ["boto3>=1.28"]

[tool.setuptools.packages.find]
include = ["tinygo*"]

[project.scripts]
tinygo = "tinygo.cli:main"
